# Use bufio in index